fixed ch 1 bug
Got all courses loaded in
