--- src/frontend/src/pages/Chapters/Chapter24/course.md
@@ -1,1 +1,0 @@
-# Chapter 24: Comming soon...
