--- src/frontend/src/pages/Chapters/Chapter24/exercise.md
@@ -1,1 +1,0 @@
-// Comming soon...
