--- src/frontend/src/pages/Chapters/Chapter24/index.ts
@@ -1,12 +1,0 @@
-/* eslint import/no-webpack-loader-syntax: off */
-// @ts-ignore
-import course from '!raw-loader!./course.md'
-/* eslint import/no-webpack-loader-syntax: off */
-// @ts-ignore
-import exercise from '!raw-loader!./exercise.md'
-/* eslint import/no-webpack-loader-syntax: off */
-// @ts-ignore
-import solution from '!raw-loader!./solution.md'
-import { Data } from 'pages/Chapter/Chapter.controller'
-
-export const data: Data = { course, exercise, solution, supports: { }, questions: [] }
