--- src/frontend/src/pages/Chapters/Chapter24/solution.md
@@ -1,1 +1,0 @@
-// Comming soon...
