+++ src/frontend/src/pages/Course/Course.controller.tsx
@@ -1,17 +1,20 @@
 import * as React from 'react'
+
 import { Data } from '../Chapter/Chapter.controller'
 import { CourseView } from './Course.view'
 
 export interface Course {
-    path: string | undefined
+    path: string
     description: string | undefined
 }
 
-export interface ChapterData {
-    pathname: string
-    name: string
-    data: Data
+export type CourseData = {
+    path: string;
+    pathname: string;
+    name: string;
+    data: Course;
 }
+
 
 export const Course = () => {
