+++ src/frontend/src/pages/Courses/introToDataDefi/Chapters/Chapter2/course.md
@@ -1,2 +1,1 @@
-# Compute To Data Chapter 2
-How to use Compute To Data
+# Chapter 2: Coming soon...
