+++ src/frontend/src/pages/Courses/introToDataDefi/Chapters/Chapter3/course.md
@@ -1,2 +1,1 @@
-# Compute To Data Chapter 1
-What is compute to data? Why is it relevant?
+# Chapter 3: Coming soon...
