+++ src/frontend/src/pages/Courses/introToDataDefi/Chapters/Chapter4/course.md
@@ -1,2 +1,1 @@
-# Compute To Data Chapter 2
-How to use Compute To Data
+# Chapter 4: Coming soon...
