+++ src/frontend/src/pages/Courses/introToDataDefi/index.ts
@@ -5,6 +5,6 @@ import data from '!raw-loader!./module.md'
 import { Course } from '../../Course/Course.controller'
 
 export const course: Course = { 
-    path: "computeToData", 
+    path: "introToDataDefi", 
     description: data 
 }
