+++ src/frontend/src/pages/Courses/ocean101/Chapters/Chapter2/course.md
@@ -1,14 +1,17 @@
 # Chapter 2: Key Concepts of Blockchain
+
 #### Difficulty: **2/5** \| Estimated reading time: **3 min**
 
 <dialog character="jellyfish">You are now below water. This is a new environment. It feels like flying, it’s awesome. Most divers will tell you that they got hooked by the experience of moving freely in 3D from their very first dive. Same here. You will get a taste of a new universe enabled by collaborative technologies, in uncharted territories.</dialog>
 
-Understanding what is a blockchain and a smart contract will help fully grasp the value that Ocean Protocol can provide. If you are already familiar with these concepts, you can go directly to [chapter 6](https://oceanacademy.io/chapter-6).
+Understanding what is a blockchain and a smart contract will help fully grasp the value that Ocean Protocol can provide. If you are already familiar with these concepts, you can go directly to <a href="https://oceanacademy.io/ocean101/chapter-6" target="_blank" >chapter 6</a>.
 
 ### A Blockchain
+
 is simply a ledger that chronologically records the facts that it is programmed to record. This ledger is duplicated across participants, and public, which allows everyone to check that its entries are correct. This removes the need for a central ‘validation authority’: you don’t need to trust when you can verify.
 
-**Blockchains work like enhanced Google sheets.** You can think of a blockchain as a Google sheet with universal read and write access, but that can only accept new inputs if those new inputs are compliant with what the sheet was programmed to accept. 
+**Blockchains work like enhanced Google sheets.** You can think of a blockchain as a Google sheet with universal read and write access, but that can only accept new inputs if those new inputs are compliant with what the sheet was programmed to accept.
+
 - Only those with a valid input can change the document
 - Old entries cannot be altered or deleted
 - Everyone can see that no one provided a non-compliant input
