+++ src/frontend/src/reducers/loginDrawer.ts
@@ -18,12 +18,10 @@ export function loginDrawer(state = loginDrawerDefaultState, action: any): Login
             return loginDrawerDefaultState
         }
         case SHOW_MENU_DRAWER:
-            console.log("loginDrawer SHOW_DRAWER\nediting showingMenu...")
             return {
                 showingMenu: true,
             }
         case HIDE_MENU_DRAWER:
-            console.log("loginDrawer HIDE_DRAWER\nediting showingMenu...")
             return {
                 showingMenu: false,
             }
